Databasse Creation
Makefile and SQL file to create a project and user table in
a SQLite3 database. It should be noted that all datatypes are
placed into an affinty of either TEXT, NUMERIC, INTEGER, REAL,
or NONE and lose all length restrictions (http://www.sqlite.org/datatype3.html),
but an attempt was made to go by MySQL datatypes in case of later
migration away from SQLite.