Upload existing file marked for deletion error
When a file on the server is marked for deletion, but not yet deleted
it prevents the user from adding a new dataset with a file of the same
name. A new check is added to instruct users to reload a project in this
situation so that the filename is no longer used.