Additional dependency checks for glyphs
-Does not allow sub-targets to be deleted if they are used by an action
or event
-Added reference to project to DependencyChecker (set in
ProjectViewModel) so anything can call to DependencyChecker, not just
ProjectViewModel or things that have references to the project.
-Updated calls to DepedencyChecker methods to remove the project
argument
-Updated USMap.subTargets getter to flatten all sub target arrays into
a 1-dimensional array in order to make calls higher up more generic and
treat sub-targets as an array. Before we just had USMap.subtargets
returning the glyph array…This isn’t right because what if we have
other sub-targets? If we return something like [[glyph array], object,
object], methods higher up won’t know what to do with the glyph array.